% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_star_interactive.R
\name{geom_star_interactive}
\alias{geom_star_interactive}
\title{Create interactive star points}
\usage{
geom_star_interactive(...)
}
\arguments{
\item{...}{see also the [geom_star()].}
}
\description{
The geometry is based on [geom_star()].
See the documentation for those functions for more details.
}
\examples{
library(ggplot2)
library(ggiraph)
p <- ggplot(iris, aes(x=Sepal.Length,
                      y=Sepal.Width,
                      fill = Species,
                      starshape = Species,
                      tooltip = Species)
     ) +
     geom_star_interactive(size=3)
girafe(ggobj=p)
}
