% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{polygens}
\alias{polygens}
\title{Create the coordinates of regular polygens}
\usage{
polygens(t, n, r)
}
\arguments{
\item{t}{the polar coordinate angle at (0, 0)}

\item{n}{the number of side of regular polygens}

\item{r}{the radius of regular polygens}
}
\description{
Create the coordinates of regular polygens
}
\keyword{internal}
