% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{show_starshapes}
\alias{show_starshapes}
\title{Show the total star shapes}
\usage{
show_starshapes(...)
}
\arguments{
\item{...}{see also \code{\link[ggplot2]{theme}}.}
}
\value{
gg object
}
\description{
Show the total star shapes
}
\examples{
p <- show_starshapes()
p
}
\author{
Shuangbin Xu
}
