% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-starshape.R
\name{starshape_pal}
\alias{starshape_pal}
\title{starshape palette (discrete)}
\usage{
starshape_pal(default = TRUE)
}
\arguments{
\item{default}{should starshapes be reorder (1, 13, 15, 11, 12, 14, 29, 2, 27) or not?}
}
\description{
starshape palette (discrete)
}
